katana-nbi/requirements.txt
Flask==1.1.1

# API.
flask-classful==0.14

PyYAML==3.13

# Application server for both development and production.
gunicorn==19.9.0

openstacksdk==0.23.0

#PyONE - OpenNebula
pyone==5.8.2

# PyMongo
pymongo==3.7.2

# Tnglib
# git+https://github.com/sonata-nfv/tng-cli

# Flask CORS
flask-cors

# Kafka
kafka-python==1.4.7